Strip out DOCTYPE declaration.
As indicated in the comments, this greatly slows down GSearch, as it
tries to download the DTD definitions from w3.org any time we load
an HOCR file... Wiping the doctype should greatly speed it up.

Ideally, a smarter entity resolver could be used, but adding it in
seems non-trivial (and requires changes to the GSearch code).